tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/cssrule_const_keyframes.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - KEYFRAMES_RULE on CSSRule interface</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#dom-interfaces-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The CSSRule interface has a constant called
   KEYFRAMES_RULE with value 7">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var log = document.querySelectorAll('p')[0].firstChild;
         var passed = false;
         try {
            var before = CSSRule.KEYFRAMES_RULE;
            CSSRule.KEYFRAMES_RULE = 1;
            var after = CSSRule.KEYFRAMES_RULE;
            var passed = (before == after && after == 7);
         } catch(e) {
         }
         test(function(){assert_true(passed);}, testName); done();
         log.data = (passed ? 'PASS' : 'FAIL');
      };
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Testing... (JavaScript required)</p>
</body>
</html>
